UBUNTU: SAUCE: AF_ECONET prevent kernel stack overflow
CVE-2010-3848

If CONFIG_ECONET_AUNUDP, the function declares a variable-length array
of iovec's on the stack:

static int econet_sendmsg(struct kiocb *iocb, struct socket *sock,
                          struct msghdr *msg, size_t len)
{
        ...
        ...
        struct iovec iov[msg->msg_iovlen+1];
        ...

"msg->msg_iovlen" is a user-controlled value <= 1024; 1024 'struct
iovec's is enough to overflow the kernel stack and clobber the
'thread_info' struct on either 32- or 64-bit systems with or without
CONFIG_4KSTACKS.

The code populates 'iov' with user-controlled values, so this is almost
certainly exploitable for privesc.

This patch has yet to be submitted upstream, hence marking it as SAUCE
for now.

OriginalAuthor: Phil Blundell <philb@gnu.org>

Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>