UBUNTU: SAUCE: AF_ECONET saddr->cookie prevent NULL pointer dereference
CVE-2010-3849

'saddr' comes from userspace and may be NULL (if non-NULL, it will have
been copied to kernelspace), leading to a NULL-pointer dereference. This
saddr->cookie dereference appears twice, once in the "hardware" case and
once in the "UDP emulation" case. This is easily exploitable for an
oops, but probably not anything more.

This patch has yet to be submitted upstream, hence marking it as SAUCE
for now.

Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>