UBUNTU: SAUCE: ensure root is ready before running usermodehelpers in it
On some machines depmod errors are emitted during boot:

    modprobe: Fatal: Could not load /lib/modules/2.6.35-22-generic/modules.dep
	no such file or directory

This occur because kernel initialisation can (rarely) require additional
modules.  This triggers a modprobe of that module via usermodhelper.
Where this occurs before the initramfs expand has completed required files
may not be found.  These appear to be benign in the majority of cases as
udev cold plug will load these modules, but it is possible that critical
modules are not loaded and boot fails.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Brad Figg <brad.figg@canonical.com>