UBUNTU: [Config] Allow insertchanges to work in later version chroots
This is a very minor change which only affects one maintenance script.
No changes to kernel code.

The command "git-log" was used in the script, which prevented the
insertchanges target from executing unless you were in a hardy chroot.

This changes that to "git log". I have confirmed that it still works in
a hardy chroot, and it also now works in a Lucid chroot.

Signed-off-by: Steve Conklin <sconklin@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>